src/main/webapp/app/entities/promotion/update/promotion-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize } from 'rxjs/operators';

import { PromotionFormService, PromotionFormGroup } from './promotion-form.service';
import { IPromotion } from '../promotion.model';
import { PromotionService } from '../service/promotion.service';
import { PromotionContentType } from 'app/entities/enumerations/promotion-content-type.model';

@Component({
  selector: 'yali-promotion-update',
  templateUrl: './promotion-update.component.html',
})
export class PromotionUpdateComponent implements OnInit {
  isSaving = false;
  promotion: IPromotion | null = null;
  promotionContentTypeValues = Object.keys(PromotionContentType);

  editForm: PromotionFormGroup = this.promotionFormService.createPromotionFormGroup();

  constructor(
    protected promotionService: PromotionService,
    protected promotionFormService: PromotionFormService,
    protected activatedRoute: ActivatedRoute
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ promotion }) => {
      this.promotion = promotion;
      if (promotion) {
        this.updateForm(promotion);
      }
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const promotion = this.promotionFormService.getPromotion(this.editForm);
    if (promotion.id !== null) {
      this.subscribeToSaveResponse(this.promotionService.update(promotion));
    } else {
      this.subscribeToSaveResponse(this.promotionService.create(promotion));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IPromotion>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(promotion: IPromotion): void {
    this.promotion = promotion;
    this.promotionFormService.resetForm(this.editForm, promotion);
  }
}
